+++ Sources/Tracking/Service/Request/ScreenViewDefaultData.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+// ScreenViewData is the standard data supplied along with an automatic screenview
+internal struct ScreenViewData: Codable {}