+++ Sources/Tracking/Service/Request/TrackRequestBody.swift
@@ -2,6 +2,7 @@ import Foundation
 
 /// https://customer.io/docs/api/#operation/track
 internal struct TrackRequestBody<T: Encodable>: Encodable {
+    let type: EventType
     let name: String
     let data: T?
     let timestamp: Date?
@@ -12,5 +13,15 @@ internal struct TrackRequestBody<T: Encodable>: Encodable {
         case name
         case data
         case timestamp
+    }
+}
+
+public enum EventType: String, Codable {
+    case event
+    case screen
+
+    enum CodingKeys: String, CodingKey {
+        case event
+        case screen
     }
 }