+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [1.0.0-alpha.22](https://github.com/customerio/customerio-ios/compare/1.0.0-alpha.21...1.0.0-alpha.22) (2021-12-16)
+
+
+### Features
+
+* subsequent identifies without passing identifier ([#96](https://github.com/customerio/customerio-ios/issues/96)) ([47d9166](https://github.com/customerio/customerio-ios/commit/47d916634a5ab34842a19074f3b3c110de4e0348))
+
+# [1.0.0-alpha.21](https://github.com/customerio/customerio-ios/compare/1.0.0-alpha.20...1.0.0-alpha.21) (2021-12-13)
+
+
+### Features
+
+* add [String:Any] support to identify & track ([#94](https://github.com/customerio/customerio-ios/issues/94)) ([5a220c4](https://github.com/customerio/customerio-ios/commit/5a220c401c1df1153cdb0cb56efe5a885dd08c68))
+
+# [1.0.0-alpha.20](https://github.com/customerio/customerio-ios/compare/1.0.0-alpha.19...1.0.0-alpha.20) (2021-12-13)
+
+
+### Features
+
+* add screen view tracking ([#82](https://github.com/customerio/customerio-ios/issues/82)) ([c2034a6](https://github.com/customerio/customerio-ios/commit/c2034a6fcbcea15716529a4091538f5bb8dc6a00))
+
 # [1.0.0-alpha.19](https://github.com/customerio/customerio-ios/compare/1.0.0-alpha.18...1.0.0-alpha.19) (2021-12-03)